Add Makefile to get `npm test` AND `npm start` to work together.
This is a super dodgy hack. The tests only work when package.json type is "module" and
the actual app only works when the type is "commonjs". I'm not sure why and I don't have the
JS expertise to fix it.

The Makefile has 2 rules, test and start, which both run npm test and start after changing the
package type using `sed` in package.json.

Super dodgy, works like a charm ¯\_(ツ)_/¯